% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_methods.R
\name{archi_aitchison_distance}
\alias{archi_aitchison_distance}
\title{Aitchison distance via robCompositions::aDist}
\usage{
archi_aitchison_distance(X)
}
\arguments{
\item{X}{Numeric matrix, samples in rows and taxa in columns.}
}
\description{
Aitchison distance via robCompositions::aDist
}
\keyword{internal}
