% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ordination.R
\name{archi_ancombc_fit}
\alias{archi_ancombc_fit}
\title{Call ancombc2 with a count matrix, or phyloseq on older ANCOMBC}
\usage{
archi_ancombc_fit(mat, group)
}
\description{
Call ancombc2 with a count matrix, or phyloseq on older ANCOMBC
}
\keyword{internal}
