% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ordination.R
\name{archi_ancombc_volcano}
\alias{archi_ancombc_volcano}
\title{ANCOM-BC2 log2 fold change for two legend groups}
\usage{
archi_ancombc_volcano(df, legend_detect)
}
\description{
ANCOM-BC2 log2 fold change for two legend groups
}
\keyword{internal}
