% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_methods.R
\name{archi_coda_mult_repl}
\alias{archi_coda_mult_repl}
\title{Martín-Fernández et al. (2003) multiplicative replacement}
\usage{
archi_coda_mult_repl(X, frac = 0.65)
}
\description{
Martín-Fernández et al. (2003) multiplicative replacement
}
\keyword{internal}
