% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_composition.R
\name{archi_composition_tree}
\alias{archi_composition_tree}
\title{Fan-tree tips and the rank used to colour them}
\usage{
archi_composition_tree(df, tax, color_rank)
}
\description{
Fan-tree tips and the rank used to colour them
}
\keyword{internal}
