% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_difftree_ggplot}
\alias{archi_difftree_ggplot}
\title{Rectangular differential tree without ggtree}
\usage{
archi_difftree_ggplot(tree, lfc)
}
\description{
Rectangular differential tree without ggtree
}
\keyword{internal}
