% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_difftree_ggtree}
\alias{archi_difftree_ggtree}
\title{Harness ggtree + geom_fruit differential tree}
\usage{
archi_difftree_ggtree(tree, lfc, layout)
}
\description{
Harness ggtree + geom_fruit differential tree
}
\keyword{internal}
