% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_difftree_heat}
\alias{archi_difftree_heat}
\title{Sample or group abundance matrix for the differential-tree heatmap}
\usage{
archi_difftree_heat(df, groups, lfc)
}
\description{
Sample or group abundance matrix for the differential-tree heatmap
}
\keyword{internal}
