% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_difftree_heat_matrix}
\alias{archi_difftree_heat_matrix}
\title{Wide taxa-by-sample matrix for \link[ggtree:gheatmap]{ggtree::gheatmap}}
\usage{
archi_difftree_heat_matrix(df, groups, lfc)
}
\description{
Wide taxa-by-sample matrix for \link[ggtree:gheatmap]{ggtree::gheatmap}
}
\keyword{internal}
