% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_difftree_metacoder}
\alias{archi_difftree_metacoder}
\title{metacoder heat tree coloured by log2 fold change}
\usage{
archi_difftree_metacoder(df, lfc, tax)
}
\description{
metacoder heat tree coloured by log2 fold change
}
\keyword{internal}
