% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_group_lfc}
\alias{archi_group_lfc}
\title{Two-group log2 fold change of mean relative abundance}
\usage{
archi_group_lfc(df, group, contrast = NULL)
}
\description{
Two-group log2 fold change of mean relative abundance
}
\keyword{internal}
