% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_heattree_metacoder}
\alias{archi_heattree_metacoder}
\title{metacoder::heat_tree from a rank table plus abundances}
\usage{
archi_heattree_metacoder(df, tax)
}
\description{
metacoder::heat_tree from a rank table plus abundances
}
\keyword{internal}
