% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ordination.R
\name{archi_legend_groups}
\alias{archi_legend_groups}
\title{Samples matching two legend patterns}
\usage{
archi_legend_groups(df, legend_detect)
}
\description{
Samples matching two legend patterns
}
\keyword{internal}
