% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_mp_diff}
\alias{archi_mp_diff}
\title{mp_diff_analysis, relaxed once if the first pass finds no biomarkers}
\usage{
archi_mp_diff(mpse, relative, group_name)
}
\description{
mp_diff_analysis, relaxed once if the first pass finds no biomarkers
}
\keyword{internal}
