% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_mpse_abundance}
\alias{archi_mpse_abundance}
\title{Rarefy when possible, then store a relative-abundance assay}
\usage{
archi_mpse_abundance(mpse)
}
\description{
Rarefy when possible, then store a relative-abundance assay
}
\keyword{internal}
