% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_mpse_taxonomy}
\alias{archi_mpse_taxonomy}
\title{Capitalised rank table indexed by taxon label}
\usage{
archi_mpse_taxonomy(taxa, tax)
}
\description{
Capitalised rank table indexed by taxon label
}
\keyword{internal}
