% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_optional}
\alias{archi_optional}
\title{Stop when an optional plotting package is missing}
\usage{
archi_optional(pkg, what)
}
\description{
Stop when an optional plotting package is missing
}
\keyword{internal}
