% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_parse_phyloseq}
\alias{archi_parse_phyloseq}
\title{parse_phyloseq, after attaching metacoder so ranks_ref is visible}
\usage{
archi_parse_phyloseq(ps, to_relative = TRUE)
}
\description{
parse_phyloseq, after attaching metacoder so ranks_ref is visible
}
\keyword{internal}
