% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_tree.R
\name{archi_phylo_from_rank_paths}
\alias{archi_phylo_from_rank_paths}
\title{Edge-list taxonomy tree from per-tip rank paths}
\usage{
archi_phylo_from_rank_paths(paths, tip_labels)
}
\description{
Edge-list taxonomy tree from per-tip rank paths
}
\keyword{internal}
