% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_tree.R
\name{archi_phylo_from_rank_table}
\alias{archi_phylo_from_rank_table}
\title{ape phylo from rank columns, keeping unary nodes}
\usage{
archi_phylo_from_rank_table(rank_df, tip_labels)
}
\description{
ape phylo from rank columns, keeping unary nodes
}
\keyword{internal}
