% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_phylo_layout}
\alias{archi_phylo_layout}
\title{Rectangular layout for a phylo object (tips evenly spaced)}
\usage{
archi_phylo_layout(tree)
}
\description{
Rectangular layout for a phylo object (tips evenly spaced)
}
\keyword{internal}
