% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_methods.R
\name{archi_phyloseq_distance}
\alias{archi_phyloseq_distance}
\title{phyloseq::distance for UniFrac, JSD and DPCoA}
\usage{
archi_phyloseq_distance(counts, method, tree = NULL)
}
\description{
phyloseq::distance for UniFrac, JSD and DPCoA
}
\keyword{internal}
