% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_plot_mp_diff}
\alias{archi_plot_mp_diff}
\title{ggtree drawing of an mp_diff_analysis taxatree}
\usage{
archi_plot_mp_diff(tree, group_name)
}
\description{
ggtree drawing of an mp_diff_analysis taxatree
}
\keyword{internal}
