% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_prepare_ggtree}
\alias{archi_prepare_ggtree}
\title{ggplot2 4 renamed is.waive; released ggtree still calls it while drawing}
\usage{
archi_prepare_ggtree()
}
\description{
ggplot2 4 renamed is.waive; released ggtree still calls it while drawing
}
\keyword{internal}
