% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_methods.R
\name{archi_replace_zeros_coda}
\alias{archi_replace_zeros_coda}
\title{Replace zeros before Aitchison distance}
\usage{
archi_replace_zeros_coda(X)
}
\description{
Replace zeros before Aitchison distance
}
\keyword{internal}
