% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_sample_group}
\alias{archi_sample_group}
\title{Map a tidy-table column onto samples}
\usage{
archi_sample_group(df, split_by, samples)
}
\description{
Map a tidy-table column onto samples
}
\keyword{internal}
