% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_set_palette}
\alias{archi_set_palette}
\title{Set1-like colours for UpSet set-size bars}
\usage{
archi_set_palette(n)
}
\description{
Set1-like colours for UpSet set-size bars
}
\keyword{internal}
