% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchiteutis-package.R
\docType{package}
\name{aRchiteutis-package}
\alias{aRchiteutis}
\alias{aRchiteutis-package}
\title{aRchiteutis: Visualise and Work with Kraken2 Metagenomics Reports}
\description{
A toolkit (formerly a loose collection of scripts, also known as 'samovar') for parsing, transforming and visualising taxonomic composition tables produced by Kraken2 (and Kaiju / Bracken) metagenomics classifiers. It provides helpers to read classifier reports into tidy data frames, reshape and trim taxa, and a family of 'df2*' plotting functions covering composition (donut, stacked bar, box), alpha and beta diversity, ordination (PCA, PCoA, t-SNE), clustering, heatmaps, correlation, chord and volcano plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dsmutin/aRchiteutis}
  \item Report bugs at \url{https://github.com/dsmutin/aRchiteutis/issues}
}

}
\author{
\strong{Maintainer}: Daniil Smutin \email{dvsmutin@gmail.com}

Authors:
\itemize{
  \item Daniil Smutin \email{dvsmutin@gmail.com}
  \item Amir Taldaev
  \item Egor Lebedev
  \item Leonid Adonin
}

}
\keyword{internal}
