% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_diversity.R
\name{df2beta_bray}
\alias{df2beta_bray}
\title{Bray-Curtis beta-diversity heatmap}
\usage{
df2beta_bray(df, ...)
}
\arguments{
\item{df}{A tidy \code{tibble} from \code{\link[=get_counts]{get_counts()}}. A \link[phyloseq:phyloseq]{phyloseq::phyloseq} object
or a plain taxa-by-sample matrix are also accepted (via
\code{as_samovar_matrix()} / \code{as_samovar_df()}).}

\item{...}{Passed to \code{\link[=df2beta]{df2beta()}}.}
}
\value{
See \code{\link[=df2beta]{df2beta()}}.
}
\description{
Thin convenience wrapper around \code{\link[=df2beta]{df2beta()}} that fixes the distance function
to Bray-Curtis.
}
\examples{
path <- system.file("extdata", package = "aRchiteutis")
df <- get_counts(path = path, pattern = "m[13][124]_", trim_char = "_")
d <- df2beta_bray(df[df$clade == "G", ], print_df = TRUE)
dim(d)

}
