% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_diversity.R
\name{df2beta_pcoa}
\alias{df2beta_pcoa}
\title{PCoA ordination of samples from a beta-diversity matrix}
\usage{
df2beta_pcoa(
  df,
  dist_function = abdiv::bray_curtis,
  treshhold_up = 1,
  treshhold_down = 0,
  add_legend = FALSE,
  add_ellipse = FALSE,
  ...
)
}
\arguments{
\item{df}{A tidy \code{tibble} from \code{\link[=get_counts]{get_counts()}}.}

\item{dist_function}{Distance function from \pkg{abdiv}
(default \link[abdiv:bray_curtis]{abdiv::bray_curtis}).}

\item{treshhold_up, treshhold_down}{Upper / lower mean-abundance thresholds
used to filter taxa.}

\item{add_legend}{Integer column index/indices (or \code{FALSE}) drawn as colour
side-bars.}

\item{add_ellipse}{Integer column index/indices (or \code{FALSE}) used to draw
grouping ellipses.}

\item{...}{Passed to \link[heatmap3:heatmap3]{heatmap3::heatmap3}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
PCoA ordination of samples from a beta-diversity matrix
}
\examples{
path <- system.file("extdata", package = "aRchiteutis")
legend <- system.file("extdata", "legend.csv", package = "aRchiteutis")
df <- get_counts(path = path, pattern = "m[13][124]_", legend = legend,
                 trim_char = "_")
df2beta_pcoa(df[df$clade == "G", ], add_legend = 7)

}
