% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_composition.R
\name{df2composition}
\alias{df2composition}
\title{Stacked bar plot of composition per sample}
\usage{
df2composition(
  df,
  order_samples = c("fpc", "hclust", "abundance", "alpha", "none")
)
}
\arguments{
\item{df}{A tidy \code{tibble} from \code{\link[=get_counts]{get_counts()}}.}

\item{order_samples}{Sample order. See Details.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Samples are ordered by the first principal component of the composition
matrix (\code{order_samples = "fpc"}). Other orders: \code{"hclust"}, \code{"abundance"},
\code{"alpha"}, \code{"none"}.
}
\examples{
path <- system.file("extdata", package = "aRchiteutis")
df <- get_counts(path = path, pattern = "m1[124]_", trim_char = "_")
df2composition(df_taxa_trim(df[df$clade != "S", ], top_taxa = 6))

}
