% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_composition.R
\name{df2donut}
\alias{df2donut}
\title{Donut plot of mean taxonomic composition}
\usage{
df2donut(df, ...)
}
\arguments{
\item{df}{A tidy \code{tibble} from \code{\link[=get_counts]{get_counts()}} (usually already trimmed to a
handful of taxa, e.g. via \code{\link[=df_taxa_trim]{df_taxa_trim()}}).}

\item{...}{Reserved for future use.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Donut plot of mean taxonomic composition
}
\examples{
path <- system.file("extdata", package = "aRchiteutis")
df <- get_counts(path = path, pattern = "m1[124]_", trim_char = "_")
df2donut(df_taxa_trim(df[df$clade != "S", ], top_taxa = 6))

}
