% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ordination.R
\name{df2pca_sp}
\alias{df2pca_sp}
\title{PCA of taxa}
\usage{
df2pca_sp(df, scale = TRUE, ...)
}
\arguments{
\item{df}{A numeric taxa-by-sample matrix, e.g. from \code{\link[=df_untidy]{df_untidy()}}.}

\item{scale}{Logical. Scale variables to unit variance before PCA.}

\item{...}{Passed to \link[factoextra:fviz_pca_var]{factoextra::fviz_pca_var}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Like \code{\link[=df2pca_sample]{df2pca_sample()}} but draws the variables (taxa) plot.
}
\examples{
path <- system.file("extdata", package = "aRchiteutis")
df <- get_counts(path = path, pattern = "m[13][124]_", trim_char = "_")
df2pca_sp(df_untidy(df, clade = "G", top = 10), scale = TRUE)

}
