% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ordination.R
\name{df2volcano}
\alias{df2volcano}
\title{Volcano plot of abundance change between two sample groups}
\usage{
df2volcano(df, legend_detect, treshhold_logAC = 0.5, treshhold_p = 0.05)
}
\arguments{
\item{df}{A tidy \code{tibble} from \code{\link[=get_counts]{get_counts()}} with legend columns.}

\item{legend_detect}{Length-2 character vector of patterns identifying the
two groups.}

\item{treshhold_logAC}{Numeric. Absolute log abundance-change threshold for
labelling.}

\item{treshhold_p}{Numeric. P-value threshold for labelling.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Runs a per-taxon t-test between the two groups defined by \code{legend_detect} and
plots the log abundance change against the t-test p-value.
}
\examples{
path <- system.file("extdata", package = "aRchiteutis")
legend <- system.file("extdata", "legend.csv", package = "aRchiteutis")
df <- get_counts(path = path, pattern = "m[1345][0-9]?_", legend = legend,
                 trim_char = "_")
df2volcano(df[df$clade == "G", ], legend_detect = c("pupa", "larvae"))

}
