% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_transforms.R
\name{df_remove_taxa}
\alias{df_remove_taxa}
\title{Remove specific taxa from the table}
\usage{
df_remove_taxa(df, taxa, clade = FALSE)
}
\arguments{
\item{df}{A tidy \code{tibble} from \code{\link[=get_counts]{get_counts()}}.}

\item{taxa}{Character vector of taxa names to remove.}

\item{clade}{Character or \code{FALSE}. If set, only remove matching taxa within
this clade.}
}
\value{
\code{df} with the requested taxa removed.
}
\description{
Remove specific taxa from the table
}
