% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_transforms.R
\name{df_tidy_drop_unclassified}
\alias{df_tidy_drop_unclassified}
\title{Drop unclassified taxa}
\usage{
df_tidy_drop_unclassified(df)
}
\arguments{
\item{df}{A tidy \code{tibble} from \code{\link[=get_counts]{get_counts()}}.}
}
\value{
\code{df} with rows whose \code{taxa} contains \code{"unclassified"} removed.
}
\description{
Drop unclassified taxa
}
