% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_tree.R
\name{parse_ncbi_taxonomy_xml}
\alias{parse_ncbi_taxonomy_xml}
\title{Parse an NCBI taxonomy efetch XML document}
\usage{
parse_ncbi_taxonomy_xml(xml)
}
\arguments{
\item{xml}{Character. Document from
\code{efetch.fcgi?db=taxonomy&retmode=xml}.}
}
\value{
A lineage data frame, one row per requested taxid.
}
\description{
Parse an NCBI taxonomy efetch XML document
}
