% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_difftree_ggtree}
\alias{archi_difftree_ggtree}
\title{ggtree differential tree with a fold-change bar at each tip}
\usage{
archi_difftree_ggtree(tree, lfc, layout)
}
\description{
ggtree differential tree with a fold-change bar at each tip
}
\keyword{internal}
