% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_harness.R
\name{archi_geom_fruit}
\alias{archi_geom_fruit}
\title{Fruit column that ggtreeExtra can construct from inside another package}
\usage{
archi_geom_fruit(data, mapping, offset, pwidth)
}
\description{
Fruit column that ggtreeExtra can construct from inside another package
}
\keyword{internal}
