% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_diversity.R
\name{df2alpha}
\alias{df2alpha}
\title{Alpha diversity for a chosen set of metrics}
\usage{
df2alpha(
  df,
  split_by = FALSE,
  add_legend = FALSE,
  alpha_function_list = list(Shannon = abdiv::shannon, Simpson = abdiv::simpson),
  ...
)
}
\arguments{
\item{df}{A tidy \code{tibble} from \code{\link[=get_counts]{get_counts()}}. Retain species-level rows for
meaningful diversity estimates.}

\item{split_by}{Integer column index (or \code{FALSE}) used to split the samples
into groups on the x-axis.}

\item{add_legend}{Integer column index/indices (or \code{FALSE}) whose values are
combined into the colour legend.}

\item{alpha_function_list}{Named list of diversity functions (defaults to
Shannon and Simpson from \pkg{abdiv}).}

\item{...}{Passed to the geoms.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
A lighter version of \code{\link[=df2alpha_summary]{df2alpha_summary()}} using a user-supplied list of
diversity functions.
}
\examples{
path <- system.file("extdata", package = "aRchiteutis")
df <- get_counts(path = path, pattern = "m1[124]_", trim_char = "_")
df2alpha(df[df$clade == "S", ])

}
