% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_composition.R
\name{df2barplot}
\alias{df2barplot}
\title{Box plot of per-sample composition across taxa}
\usage{
df2barplot(df, ...)
}
\arguments{
\item{df}{A tidy \code{tibble} from \code{\link[=get_counts]{get_counts()}}.}

\item{...}{Reserved for future use.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Box plot of per-sample composition across taxa
}
\examples{
path <- system.file("extdata", package = "aRchiteutis")
df <- get_counts(path = path, pattern = "m1[124]_", trim_char = "_")
df2barplot(df_taxa_trim(df[df$clade != "S", ], top_taxa = 6))

}
