% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_composition.R
\name{df2composition}
\alias{df2composition}
\title{Stacked bar plot of composition per sample}
\usage{
df2composition(df)
}
\arguments{
\item{df}{A tidy \code{tibble} from \code{\link[=get_counts]{get_counts()}}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Stacked bar plot of composition per sample
}
\examples{
path <- system.file("extdata", package = "aRchiteutis")
df <- get_counts(path = path, pattern = "m1[124]_", trim_char = "_")
df2composition(df_taxa_trim(df[df$clade != "S", ], top_taxa = 6))

}
