% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_ordination.R
\name{df2corrplot}
\alias{df2corrplot}
\title{Correlation plot between taxa}
\usage{
df2corrplot(df, k_means = FALSE, ...)
}
\arguments{
\item{df}{A numeric taxa-by-sample matrix, e.g. from \code{\link[=df_untidy]{df_untidy()}}. A
\link[phyloseq:phyloseq]{phyloseq::phyloseq} object or a long \code{df} from \code{\link[=get_counts]{get_counts()}} is also
accepted and reduced to a genus-level matrix first (a plain matrix is used
as-is).}

\item{k_means}{Integer or \code{FALSE}. Number of clusters to outline with
rectangles.}

\item{...}{Passed to \link[corrplot:corrplot]{corrplot::corrplot}.}
}
\value{
Invisibly, the correlation matrix (a plot is drawn as a side effect).
}
\description{
Correlation plot between taxa
}
\examples{
path <- system.file("extdata", package = "aRchiteutis")
df <- get_counts(path = path, pattern = "m[13][124]_", trim_char = "_")
mat <- df_untidy(df, clade = "G", top = 12, scale = "scale")
grDevices::pdf(tempfile()); df2corrplot(mat, k_means = 3)
grDevices::dev.off()

}
